feat: 냉장고 정보 입력
