+++ package.json
@@ -9,7 +9,7 @@
     "web": "expo start --web"
   },
   "dependencies": {
-    "@react-native-community/datetimepicker": "6.7.3",
+    "@react-native-community/datetimepicker": "^7.1.0",
     "@react-native-picker/picker": "2.4.8",
     "@react-navigation/bottom-tabs": "^6.5.7",
     "@react-navigation/material-top-tabs": "^6.6.2",
@@ -30,6 +30,7 @@
     "react": "18.2.0",
     "react-native": "0.71.7",
     "react-native-chart-kit": "^6.12.0",
+    "react-native-datepicker": "^1.7.2",
     "react-native-linear-gradient": "^2.6.2",
     "react-native-modal-datetime-picker": "^15.0.1",
     "react-native-pager-view": "6.1.2",
