+++ src/navigations/RefrigeratorStackNavigation.js
@@ -1,7 +1,7 @@
 import { createNativeStackNavigator } from "@react-navigation/native-stack";
 import React from "react";
 import MyRefrigeratorMain from "../screen/myrefrigerator/MyRefrigeratorMain";
-// import BarcodeScanner from "../screen/myrefrigerator/BarcordScanner";
+// import BarcodeScanner from "../screen/myrefrigerator/BarcodeScanner";
 import MyRefrigeratorResult from "../screen/myrefrigerator/MyRefrigeratorResult";
 
 
@@ -17,7 +17,7 @@ const RefrigeratorStackNavigation = () => {
       }}
     >
       <Stack.Screen name="MyRefrigeratorMain" component={MyRefrigeratorMain} />
-      {/* <Stack.Screen name="BarcordScanner" component={BarcodeScanner} /> */}
+      {/* <Stack.Screen name="BarcodeScanner" component={BarcodeScanner} /> */}
       <Stack.Screen name="MyRefrigeratorResult" component={MyRefrigeratorResult} />
     </Stack.Navigator>
   );
