+++ src/navigations/RootStackNavigation.js
@@ -6,8 +6,8 @@ import SignStackNavigation from "./SignStackNavigation";
 import CameraScreen from '../screen/ripeness/CameraScreen';
 import GroupBuyingPostScreen from '../screen/groupbuying/GroupBuyingPostScreen';
 import LoadingScreen from '../screen/prediction/Loading';
-import BarcodeScanner from "../screen/myrefrigerator/BarcordScanner";
-// import BarcodeTest from "../screen/myrefrigerator/BarcordTest";
+import BarcodeScanner from "../screen/myrefrigerator/BarcodeScanner";
+// import BarcodeTest from "../screen/myrefrigerator/BarcodeTest";
 
 const Stack = createNativeStackNavigator();
 
@@ -24,7 +24,7 @@ const RootStackNavigation = () => {
       <Stack.Screen name="Sign" component={SignStackNavigation} />
       <Stack.Screen name="Memo" component={MemoScreen} /> 
       <Stack.Screen name="Camera" component={CameraScreen}/>
-      <Stack.Screen name="BarcordScanner" component={BarcodeScanner} />
+      <Stack.Screen name="BarcodeScanner" component={BarcodeScanner} />
       {/* <Stack.Screen name="BarcordTest" component={BarcodeTest} /> */}
       <Stack.Screen name="GBPost" component={GroupBuyingPostScreen}/>
       <Stack.Screen name="Loading" component={LoadingScreen} />
